Complete bloglist authentication and authorization

+++ part4/bloglist/controllers/blogs.js
@@ -1,22 +1,41 @@
 const blogsRouter = require('express').Router()
+
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user')
 
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const blog = new Blog({
+    ...request.body,
+    user: request.user._id
+  })
 
   const savedBlog = await blog.save()
 
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
   const id = request.params.id
+
+  const blog = await Blog.findById(id)
+
+  if (!blog) {
+    return response.status(404).json({
+      error: 'blog not found'
+    })
+  }
+
+  if (blog.user.toString() !== request.user._id.toString()) {
+    return response.status(403).json({
+      error: 'user not authorized to delete this blog'
+    })
+  }
 
   await Blog.findByIdAndDelete(id)
